updated user detail view, user password recovery and pass change

+++ users/views.py
@@ -6,7 +6,7 @@ from rest_framework.exceptions import ValidationError, PermissionDenied
 from listings.imagekit import create_image, destroy_image
 from common.mail_services import send_email
 from urllib.parse import parse_qs
-from random import randint
+from .verification_code_generator import generate_verification_code
 from django.utils import timezone
 from rest_framework.views import APIView
 from rest_framework.response import Response
@@ -39,7 +39,17 @@ class UserDetailView(
     queryset = User.objects.all()
     serializer_class = UserSerializer
     
+    def perform_update(self, serializer):
+        user = serializer.instance
+        if user != self.request.user and not self.request.user.is_staff:
+            raise PermissionDenied("You can only edit your own profile")
+        
+        serializer.save()
+    
     def perform_destroy(self,instance):
+        if self.request.user != instance and not self.request.user.is_staff: 
+            raise PermissionDenied("You can only delete your own profile!")
+        
         if instance.storage_key:
             destroy_image(instance.storage_key)
         
@@ -48,41 +58,39 @@ class UserDetailView(
 
 
 class SendEmailVerficationCode(
-     # UserPermission,
+     UserPermission,
     APIView):
 
     
     def post(self,request):
         user = request.user
 
-        if not user.is_authenticated:
-            return Response({"error":"You are not logged in"}, status=400)
             
         if user.is_verified:
             return Response({"error":"You are already verified"}, status=400)
         
         
-        verification_number = randint(100000, 999999)
-        user.email_verification_code = verification_number
+        verification_code = generate_verification_code()
+        user.email_verification_code = verification_code
         user.email_verification_code_date = timezone.now()
         user.save(update_fields=["email_verification_code","email_verification_code_date"])  
 
         full_name =f"{user.first_name} {user.last_name}"
         
-        send_email(full_name,user.email,"Verificaition code", f"Your verification code is here: {verification_number}")
+        send_email(full_name,user.email,"Verificaition code", f"Your verification code is here: {verification_code}")
         
         return Response({"message":"Verification code has been sent to email"})
 
         
         
 class VerifyUser(
-         # UserPermission,
+         UserPermission,
          APIView):
     
     def post(self,request):
         user = request.user
         
-        if not user.is_authenticated:
+        if not user.is_authenticated or not user.is_staff:
             return Response({"error":"You are not logged in"}, status=400)
             
         if not user.email_verification_code_date or not user.email_verification_code:
@@ -112,7 +120,99 @@ class VerifyUser(
         user.save(update_fields=["is_verified","email_verification_code","email_verification_code_date"])
         
         return Response({"message":"Succeefully verified User!"})
-   
+    
+
+class SendPasswordRecoveryEmail(
+    APIView):
+    def post(self,request):
+        email = request.data.get("email")
+        
+        if not email:
+            return Response({"error":"Email is required!"}, status=400)
+        
+
+        user = User.objects.filter(email=email).first()
+        
+        if not user:
+            return Response({"error":"No email with this address was found"}, status=404)
+        
+        time_now = timezone.now()
+        verification_code = generate_verification_code()
+
+        user.password_recovery_code = verification_code
+        user.password_recovery_code_date = time_now
+        
+        user.save(update_fields=["password_recovery_code","password_recovery_code_date"])
+        
+        full_name =f"{user.first_name} {user.last_name}"
+        
+        send_email(full_name,user.email,"Password recovery code", f"Your Password recovery code is here: {verification_code}")
+        
+        return Response({"message":"Password recovery code has been sent to email"})
+
+class RecoverPassword(
+    APIView):
+    def post(self,request):
+        email = request.data.get("email")
+        new_password = request.data.get("new_password")
+        code = self.request.data.get("code")
+        
+        if not email or not new_password or not code:
+            return Response({"error":"Email, new password and recovery code are required!"}, status=400)
+        
+        user = User.objects.filter(email=email).first()
+        
+        if not user:
+            return Response({"error":"User not found"}, status=404)
+        
+        if not user.password_recovery_code or not user.password_recovery_code_date:
+            return Response({"error":"You did not ask for recovery code"}, status=400)
+        
+        if not code:
+            return Response({"error":"recovery code is required!"}, status=400)
+            
+        time_now = timezone.now()
+        time_difference = time_now - user.password_recovery_code_date
+        
+        if time_difference.total_seconds() > 600:
+            return Response({"error":"Recovery code expired!"}, status=400)
+        
+        if str(code) != str(user.password_recovery_code):
+            return Response({"error":"wrong recovery code"}, status=400)
+
+        
+        user.password_recovery_code = None
+        user.password_recovery_code_date = None
+        user.set_password(new_password)
+        
+        user.save(update_fields=["password_recovery_code","password_recovery_code_date","password"])
+        
+        return Response({"message":"Succeefully Recovevered your password!"})
+        
+        
+class ChangePassword(
+    UserPermission,
+    APIView):
+    
+    def post(self,request):
+        current_password = request.data.get("current_password")
+        new_password = request.data.get("new_password")
+        
+        user = request.user
+        
+        if not current_password or not new_password:
+            return Response({"error":"Both fields need to be provided!"}, status=400)
+        
+        if not user.check_password(current_password):
+            return Response({"error":"password is wrong!"},status=400)
+        
+        user.set_password(new_password)
+        
+        user.save(update_fields=["password"])
+        
+        return Response({"message":"Your password has been successfully changed!"})
+        
+        
 class UserMeView(
      # UserPermission,
     generics.RetrieveAPIView):
